requirements.txt
flask
tensorflow
dlib
opencv-python
pillow